requirements.txt
Django==3.2.19
django-environ==0.8.1
django-cors-headers==3.12.0
django_extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-storages[google]==1.13.2
psycopg2-binary==2.9.5
gunicorn==20.1.0
google-cloud-secret-manager==2.15.1
google-cloud-storage==2.7.0